quilkin_xdp
Struct EbpfProgram

pub struct EbpfProgram {
    pub external_port: NetworkU16,
    pub qcmp_port: NetworkU16,
    /* private fields */
}

Fields
external_port: NetworkU16

The external port is a variable that we modify at load time so the eBPF program can filter out which packets it is interested in. This needs to be the same port used in the I/O loop to determine if the packet is sent from a client or a server

qcmp_port: NetworkU16

The port QCMP packets are sent to

Implementations

impl EbpfProgram

pub fn load(external_port: u16, qcmp_port: u16) -> Result<Self, LoadError>

Loads the XDP program.

The external port, the port used by clients, must be passed in due to how globals work in eBPF.

pub fn create_and_bind_sockets( &mut self, nic: NicIndex, umem_cfg: UmemCfg, device_caps: &NetdevCapabilities, ring_cfg: RingConfig, ) -> Result<Vec<XdpWorker>, BindError>

Creates and binds sockets

pub fn attach( &mut self, nic: NicIndex, flags: XdpFlags, ) -> Result<XdpLinkId, ProgramError>

pub fn detach(&mut self, link_id: XdpLinkId) -> Result<(), ProgramError>